Implement LED/buzzer bring-up with MISRA-style comments and CubeIDE guide.
Stay on UI only: self-test red/yellow/green then a short beep, then green
heartbeat. App no longer touches power pins. Beginner CubeIDE walkthrough
compares each click with Arduino and explains why.

+++ Firmware/Rtos/Src/freertos_hooks.c
@@ -1,10 +1,16 @@
+/**
+ * @file    freertos_hooks.c
+ * @brief   حافظه Idle Task وقتی Static Allocation روشن است.
+ *
+ * FreeRTOS علاوه بر Task تو، یک Task داخلی Idle دارد.
+ * اگر malloc ممنوع باشد، باید بافر Idle را هم خودت بدهی.
+ *
+ * اگر CubeMX همین توابع را ساخت و لینکر گفت duplicate،
+ * این فایل را از Build خارج کن. توضیح در docs/01-led-buzzer-cubeide.md
+ */
+
 #include "FreeRTOS.h"
 #include "task.h"
-
-/**
- * Required when configSUPPORT_STATIC_ALLOCATION = 1.
- * If CubeMX already generated these functions, exclude this file from the build.
- */
 
 static StaticTask_t s_idle_tcb;
 static StackType_t s_idle_stack[configMINIMAL_STACK_SIZE];
@@ -15,7 +21,7 @@ void vApplicationGetIdleTaskMemory(StaticTask_t **ppxIdleTaskTCBBuffer,
 {
     *ppxIdleTaskTCBBuffer = &s_idle_tcb;
     *ppxIdleTaskStackBuffer = s_idle_stack;
-    *pulIdleTaskStackSize = configMINIMAL_STACK_SIZE;
+    *pulIdleTaskStackSize = (uint32_t)configMINIMAL_STACK_SIZE;
 }
 
 #if (configUSE_TIMERS == 1)
@@ -28,7 +34,7 @@ void vApplicationGetTimerTaskMemory(StaticTask_t **ppxTimerTaskTCBBuffer,
 {
     *ppxTimerTaskTCBBuffer = &s_timer_tcb;
     *ppxTimerTaskStackBuffer = s_timer_stack;
-    *pulTimerTaskStackSize = configTIMER_TASK_STACK_DEPTH;
+    *pulTimerTaskStackSize = (uint32_t)configTIMER_TASK_STACK_DEPTH;
 }
 #endif
 
@@ -36,6 +42,8 @@ void vApplicationStackOverflowHook(TaskHandle_t xTask, char *pcTaskName)
 {
     (void)xTask;
     (void)pcTaskName;
+
+    /* استک Task پر شده. در LED اگر این‌جا آمدی TASK_STACK_UI را زیاد کن. */
     taskDISABLE_INTERRUPTS();
     for (;;)
     {
